+++ packages/ihris/src/index.js
@@ -2,4 +2,5 @@ import * as Adaptor from './Adaptor.js';
 export default Adaptor;
 
 export * from './Adaptor.js';
+export * as fhir from './fhir.js';
 export * as http from './http.js';