+++ packages/ihris/test/helpers.js
@@ -1,0 +1,3 @@
+import { enableMockClient } from '@openfn/language-common/util';
+
+export const testServer = enableMockClient('https://ihris.example.com');